Create repository.

+++ src/Application.java
@@ -5,11 +5,11 @@ class Application {
     private static Object[] parameters() {
 //        byte prettiness, String name, byte weight, String meow
         Object[] parameters = new Object[4];
-        parameters[0] = (byte)(Math.random()*100);
+        parameters[0] = (int)(Math.random()*100);
         String[] names = {"Oscar", "Max", "Bella", "Tiger", "Molly", "Sam", "Max", "Misty",
                 "Coco", "Simba", "Milo", "Angel", "Chloe", "Lucy", "Missy", "Sacha", "Lala"};
         parameters[1] = names[(int)(Math.random()*names.length)];
-        parameters[2] = (byte)(Math.random()*11);
+        parameters[2] = (int)(Math.random()*11);
         parameters[3] = "Meow";
         for (int i = 0; i < (int)(Math.random()*7); ++i) {
             parameters[3] += " meow";
@@ -21,10 +21,10 @@ class Application {
 
     public static void main(String[] args) {
         Object[] params = parameters();
-        Kotik catFirst = new Kotik((byte)params[0], (String)params[1], (byte)params[2], (String)params[3]);
+        Kotik catFirst = new Kotik((int)params[0], (String)params[1], (int)params[2], (String)params[3]);
         Kotik catSecond = new Kotik();
         params = parameters();
-        catSecond.setKotik((byte)params[0], (String)params[1], (byte)params[2], (String)params[3]);
+        catSecond.setKotik((int)params[0], (String)params[1], (int)params[2], (String)params[3]);
 
         catFirst.liveAnotherDay();
